requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.1
anaconda==0.0.1.1
anaconda-client==1.2.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
blobfile==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
clyent==1.2.1
dj-database-url==1.2.0
Django==4.1.7
django-heroku==0.3.1
filelock==3.9.0
Flask==2.2.3
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
Navigator==0.2.2
openai==0.27.0
pycryptodomex==3.17
pyee==9.0.4
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
slack-bolt==1.16.2
slack-sdk==3.20.0
slackeventsapi==3.0.1
SQLAlchemy==2.0.5.post1
sqlparse==0.4.3
tiktoken==0.3.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
whitenoise==6.4.0
yarl==1.8.2
zipp==3.13.0
